Refine ARC-AGI scraper
