+++ package.json
@@ -12,8 +12,7 @@
     "scrape:livebench": "tsx scripts/scrape_livebench.ts",
     "scrape:simplebench": "tsx scripts/scrape_simplebench.ts",
     "scrape:lmarena-text": "tsx scripts/scrape_lmarena_text.ts",
-    "scrape:arc-agi-1": "tsx scripts/scrape_arc_agi_1.ts",
-    "scrape:arc-agi-2": "tsx scripts/scrape_arc_agi_2.ts",
+    "scrape:arc-agi": "tsx scripts/scrape_arc_agi.ts",
     "scrape:aider-polyglot": "tsx scripts/scrape_aider-polyglot.ts",
     "scrape:hle": "tsx scripts/scrape_hle.ts",
     "scrape:gpqa-diamond": "tsx scripts/scrape_gpqa_diamond.ts",
